changed int to double in DayScore

+++ src/model/DayScore.java
@@ -14,7 +14,7 @@ public class DayScore {
     // MODIFIES: this
     // EFFECTS: sets the dayScore to the given value.
     //          if given value is not within in [0, 10], then throw new dayScoreOutOfBoundsException
-    public void setDayScore(int dayScore) throws dayScoreOutOfBoundsException{
+    public void setDayScore(double dayScore) throws dayScoreOutOfBoundsException{
         if (dayScore >= 0 && dayScore <= 10) {
             DayScore.dayScore = dayScore;
         }
@@ -24,7 +24,7 @@ public class DayScore {
         }
     }
 
-    public int getDayScore(){
+    public double getDayScore(){
         return dayScore;
     }
